Change docs code-snippet types to tsx
The tanstack.com site shows the code-snippets as `bash` as a fallback,
as js is not expected as a language type. (see
[here](https://github.com/TanStack/tanstack.com/blob/main/app/components/CodeBlock.tsx)
for allowed language types).

In addition to allowing JS as a snippet type (https://github.com/TanStack/tanstack.com/pull/17),
changing the documentation to be `tsx` ensures that syntax
highlighting and the language type appears correct.